% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intdct.R
\name{intDCT}
\alias{intDCT}
\title{intDCT: the Integer DCT}
\usage{
intDCT(x, Ch = NULL)
}
\arguments{
\item{x}{the input integer vector}

\item{Ch}{optional DCT-IV matrix, which should be half size of x}
}
\value{
transformed integer vector
}
\description{
intDCT: the Integer DCT
}
\examples{
\dontrun{
library(RintDCT)
x <- floor(runif(100)*100)
Ch <- DCT4(50)
y <- intDCT(x,Ch)
}

}
