% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intdct.R
\name{intIDCT}
\alias{intIDCT}
\title{intIDCT: the Integer Inverse DCT}
\usage{
intIDCT(y, Ch = NULL)
}
\arguments{
\item{y}{the input integer vector}

\item{Ch}{optional DCT-IV matrix, which should be half size of x}
}
\value{
transformed integer vector
}
\description{
intIDCT: the Integer Inverse DCT
}
\examples{
\dontrun{
library(RintDCT)
y <- floor(runif(100)*100)
Ch <- DCT4(50)
x <- intDCT(y,Ch)
}

}
